build: remove superfluous+overspecific check for libguile-2.0
Guile 2.2 was released over 2 years ago
and has since been repackaged and released by many distributors,
so we should support building with that.

Using AC_CHECK_LIB for this at all turns out to be redundant
because GUILE_FLAGS already implicitly checks
for usable guile libraries (via pkg-config),
and also turns out to be over-constraining
because GUILE_FLAGS invokes GUILE_PKG which in turn
checks for any of several guile versions
(as of guile 2.2, that list of versions is [2.2 2.0 1.8]).

Fixes: <https://savannah.gnu.org/bugs/?57035>